+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Refactor) Anonymous inspector functions
 - (Feature) Recursive OwnerReference discovery
 - (Maintenance) Add check make targets
+- (Feature) Create support for local variables in actions.
 
 ## [1.2.11](https://github.com/arangodb/kube-arangodb/tree/1.2.11) (2022-04-30)
 - (Bugfix) Orphan PVC are not removed